# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs API
  contact: {}
  version: '1.0'
servers:
- url: https://starbugz-jobs-api.herokuapp.com/api/v1
  variables: {}
paths:
...
  /jobs:
    post:
      tags:
      - Jobs
      summary: create a job
      operationId: createajob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/createajobrequest'
              - example:
                  company: 网易
                  position: 游戏技术美术
            example:
              company: 网易
              position: 游戏技术美术
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Jobs
      summary: get all jobs
      operationId: getalljobs
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    createajobrequest:
      title: createajobrequest
      required:
      - company
      - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: 网易
        position: 游戏技术美术
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...